requirements.txt
streamlit==1.37.1
chromadb
sentence-transformers
numpy>=1.24,<3
pandas>=2.0,<3
pypdf
python-docx
openai==1.44.0
requests==2.32.3
python-dotenv==1.0.1
rapidfuzz==3.9.6
uvicorn==0.30.6
python-pptx
beautifulsoup4
Pillow
pytesseract
torch
open_clip_torch
lxml
pymupdf


